Extract browser probe runner
