+++ packages/runtime-playground/src/browser-command-runners.ts
@@ -24,6 +24,8 @@ import { phpBrowserWordPressDiagnosticsPlugin } from "./php-snippets.js"
 import { assertPlaygroundResponseOk, type PlaygroundRunResponse } from "./playground-command-errors.js"
 import type { PlaygroundCliServer } from "./preview-server.js"
 import type { Page } from "playwright"
+import { runBrowserProbeCommand, type BrowserProbeRunPlan } from "./browser-probe-runner.js"
+import { browserActionTargetUrls, browserAuthRequest, browserProbeWaterfallArtifact, browserRedirectDiagnosticsArtifact, browserStorageStateAuthSummary, browserStorageStateImportFromArgs, browserWordPressDiagnosticsArtifact, createBrowserProbeProgressTracker, fileSha256, installBrowserWordPressDiagnostics, installWordPressAdminAuthCookies, livenessRemainingWallTimeMs, normalizeBrowserProbeScriptCheckpoint, type BrowserCommandProgressEvent, type BrowserStorageStateImport } from "./browser-probe-support.js"
 
 const BROWSER_STEP_DEFAULT_TIMEOUT_MS = 15_000
 const BROWSER_SCRIPT_DEFAULT_TIMEOUT_MS = 120_000
@@ -31,26 +33,6 @@ const EDITOR_CANVAS_DEFAULT_IFRAME_SELECTOR = 'iframe[name="editor-canvas"]'
 const EDITOR_CANVAS_DEFAULT_LAYOUT_SELECTOR = ".block-editor-block-list__layout"
 const EDITOR_CANVAS_DEFAULT_BLOCK_SELECTOR = ".block-editor-block-list__block, [data-block]"
 const EDITOR_CANVAS_DEFAULT_TIMEOUT_MS = 30_000
-const BROWSER_PROBE_PROFILE_OVERRIDES = new Set(["browser", "device", "locale", "permissions", "throttle", "timezone", "user-agent", "viewport"])
-
-interface BrowserProbeRunPlan {
-  url: string
-  capture: Set<string>
-  waitFor: string
-  durationMs: number
-  requestedViewport?: { width: number; height: number }
-  throttleProfile?: BrowserProbeThrottleProfileDefinition
-  requestedContext: BrowserProbeContextDetails["requested"]
-  prePageScript?: string
-  script?: string
-  authRequest?: { userId: number }
-  storageStateImport?: BrowserStorageStateImport
-  failFast: boolean
-  stallTimeoutMs: number
-  wallTimeoutMs: number
-  lifecycleSelectors: string[]
-  assertions: ReturnType<typeof browserProbeAssertionsFromArgs>
-}
 
 interface BrowserActionsRunPlan {
   initialUrl?: string
@@ -65,11 +47,6 @@ interface BrowserActionsRunPlan {
   maxDomSnapshotElements: number
 }
 
-interface BrowserStorageStateImport {
-  storageState: BrowserAuthStorageState
-  summary: BrowserStorageStateImportSummary
-}
-
 interface BrowserRunPlan {
   profile: string
   capture: string[]
@@ -77,1521 +54,11 @@ interface BrowserRunPlan {
   actions?: BrowserActionsRunPlan
 }
 
-interface BrowserCommandProgressEvent {
-  command: string
-  phase: "checkpoint"
-  checkpoint: BrowserProbeScriptCheckpoint
-  progress: ReturnType<ReturnType<typeof createBrowserProbeProgressTracker>["summary"]>
-}
-
-interface BrowserProbeScriptCheckpoint {
-  name: string
-  metadata?: unknown
-  timestamp: string
-}
-
-
-
-interface BrowserProbeThrottleProfileDefinition {
-  id: string
-  cpuSlowdownRate: number
-  network: {
-    offline: boolean
-    latencyMs: number
-    downloadThroughputBytesPerSecond: number
-    uploadThroughputBytesPerSecond: number
-  }
-}
-
-const BROWSER_PROBE_THROTTLE_PROFILES: Record<string, BrowserProbeThrottleProfileDefinition> = {
-  "low-end-mobile-slow-4g": {
-    id: "low-end-mobile-slow-4g",
-    cpuSlowdownRate: 4,
-    network: {
-      offline: false,
-      latencyMs: 150,
-      downloadThroughputBytesPerSecond: 1_600_000 / 8,
-      uploadThroughputBytesPerSecond: 750_000 / 8,
-    },
-  },
-}
-
 export { BrowserCommandArtifactError, isBrowserCommandArtifactError }
+export { runBrowserProbeCommand, runSingleBrowserProbeCommand, type BrowserProbeRunPlan } from "./browser-probe-runner.js"
+export { wordpressAdminAuthCookiePhpCode } from "./browser-probe-support.js"
 export { runVisualCompareCommand } from "./browser-visual-compare.js"
 
-export async function runBrowserProbeCommand({
-  abortSignal,
-  artifactRoot,
-  command = "wordpress.browser-probe",
-  plan,
-  runtimeSpec,
-  runPlaygroundCommand,
-  server,
-  spec,
-  onProgress,
-}: {
-  abortSignal?: AbortSignal
-  artifactRoot: string
-  command?: string
-  plan?: BrowserProbeRunPlan
-  runtimeSpec?: RuntimeCreateSpec
-  runPlaygroundCommand?: (command: string, server: PlaygroundCliServer, options: { code: string } | { scriptPath: string }) => Promise<PlaygroundRunResponse>
-  server: PlaygroundCliServer
-  spec: ExecutionSpec
-  onProgress?: (event: BrowserCommandProgressEvent) => void
-}): Promise<{ artifact: BrowserProbeArtifact; artifacts?: BrowserProbeArtifact[]; output: string }> {
-  if (plan) {
-    return runSingleBrowserProbeCommand({ abortSignal, artifactRoot, command, plan, runtimeSpec, runPlaygroundCommand, server, spec, browserFilesDirectory: "files/browser", onProgress })
-  }
-
-  const profileIds = browserProbeProfileIds(spec.args ?? [])
-  if (profileIds.length === 0) {
-    const profileId = argValue(spec.args ?? [], "profile")?.trim()
-    if (profileId) {
-      const profile = browserProbeProfile(profileId)
-      return runSingleBrowserProbeCommand({
-        abortSignal,
-        artifactRoot,
-        command,
-        runtimeSpec,
-        runPlaygroundCommand,
-        server,
-        spec: { ...spec, args: browserProbeProfileArgs(spec.args ?? [], profile) },
-        browserFilesDirectory: "files/browser",
-        profileId: profile.id,
-        onProgress,
-      })
-    }
-    return runSingleBrowserProbeCommand({ abortSignal, artifactRoot, command, runtimeSpec, runPlaygroundCommand, server, spec, browserFilesDirectory: "files/browser", onProgress })
-  }
-
-  const profiles = profileIds.map((profileId) => browserProbeProfile(profileId))
-  const artifacts: BrowserProbeArtifact[] = []
-  const outputs: unknown[] = []
-  for (const profile of profiles) {
-    const result = await runSingleBrowserProbeCommand({
-      abortSignal,
-      artifactRoot,
-      command,
-      runtimeSpec,
-      runPlaygroundCommand,
-      server,
-      spec: {
-        ...spec,
-        args: browserProbeProfileArgs(spec.args ?? [], profile),
-      },
-      browserFilesDirectory: `files/browser/${profile.id}`,
-      profileId: profile.id,
-      onProgress,
-    })
-    artifacts.push(result.artifact)
-    outputs.push(JSON.parse(result.output))
-  }
-
-  const artifact = artifacts[0]
-  if (!artifact) {
-    throw new Error("wordpress.browser-probe profiles requires at least one profile")
-  }
-
-  return {
-    artifact,
-    artifacts,
-    output: `${JSON.stringify({
-      command,
-      schema: "wp-codebox/browser-probe-profile-matrix/v1",
-      profiles: outputs,
-    }, null, 2)}\n`,
-  }
-}
-
-async function runSingleBrowserProbeCommand({
-  abortSignal,
-  artifactRoot,
-  command,
-  plan,
-  runtimeSpec,
-  runPlaygroundCommand,
-  server,
-  spec,
-  browserFilesDirectory,
-  profileId,
-  onProgress,
-}: {
-  abortSignal?: AbortSignal
-  artifactRoot: string
-  command: string
-  plan?: BrowserProbeRunPlan
-  runtimeSpec?: RuntimeCreateSpec
-  runPlaygroundCommand?: (command: string, server: PlaygroundCliServer, options: { code: string } | { scriptPath: string }) => Promise<PlaygroundRunResponse>
-  server: PlaygroundCliServer
-  spec: ExecutionSpec
-  browserFilesDirectory: string
-  profileId?: string
-  onProgress?: (event: BrowserCommandProgressEvent) => void
-}): Promise<{ artifact: BrowserProbeArtifact; output: string }> {
-  const args = spec.args ?? []
-  const runPlan = plan ?? await browserProbeRunPlanFromArgs(args, profileId)
-  if (!runPlan.url) {
-    throw new Error("wordpress.browser-probe requires url=<path-or-url>")
-  }
-
-  const capture = runPlan.capture
-
-  for (const item of capture) {
-    if (!(BROWSER_PROBE_CAPTURE_VALUES as readonly string[]).includes(item)) {
-      throw new Error(`wordpress.browser-probe capture supports ${BROWSER_PROBE_CAPTURE_VALUES.join(", ")}: ${item}`)
-    }
-  }
-
-  const waitFor = runPlan.waitFor
-  const durationMs = runPlan.durationMs
-  const requestedViewport = runPlan.requestedViewport
-  const throttleProfile = runPlan.throttleProfile
-  const requestedContext = runPlan.requestedContext
-  const prePageScript = runPlan.prePageScript
-  const script = runPlan.script
-  const authRequest = runPlan.authRequest
-  const storageStateImport = runPlan.storageStateImport
-  if (authRequest && storageStateImport) {
-    throw new Error(`${command} supports one browser authentication source at a time: use auth=wordpress-admin or storage-state, not both`)
-  }
-  const failFast = runPlan.failFast
-  const stallTimeoutMs = runPlan.stallTimeoutMs
-  const wallTimeoutMs = runPlan.wallTimeoutMs
-  const livenessPolicy = browserCommandLivenessPolicy({ wallTimeoutMs, idleTimeoutMs: stallTimeoutMs })
-  const lifecycleSelectors = runPlan.lifecycleSelectors
-  const assertions = runPlan.assertions
-  const capturesConsoleForAssertions = assertions.some((assertion) => assertion.type === "no-console-errors" || assertion.type === "no-errors")
-  const capturesErrorsForAssertions = assertions.some((assertion) => assertion.type === "no-page-errors" || assertion.type === "no-errors")
-  const capturesNetworkForAssertions = browserProbeAssertionsNeedNetwork(assertions)
-  const capturesBrowserMetrics = capture.has("performance") || capture.has("memory") || browserProbeAssertionsNeedMetrics(assertions)
-  const prePageScriptMetadata = prePageScript ? browserProbeScriptMetadata(prePageScript) : undefined
-  const topology = browserPreviewTopology(args, runtimeSpec, server.serverUrl)
-  const { preview, networkPolicy } = topology
-  const routeTracker = createBrowserPreviewRouteTracker()
-  const targetUrl = topology.resolveUrl(runPlan.url)
-  const artifactSession = new BrowserArtifactSession(artifactRoot, browserFilesDirectory, { source: command, operation: "browser-probe" })
-
-  const consoleMessages: Record<string, unknown>[] = []
-  const errors: BrowserProbeErrorRecord[] = []
-  const network: BrowserProbeNetworkRecord[] = []
-  const networkTasks: Array<Promise<void>> = []
-  const checkpoints: BrowserProbeCheckpointRecord[] = []
-  const screenshotPath = artifactSession.absolutePath("screenshot.png")
-  const startedAt = now()
-  const startedAtMs = Date.now()
-  const progress = createBrowserProbeProgressTracker(startedAt, stallTimeoutMs)
-  const { devices } = await import("playwright")
-  if (requestedContext.browser && !(BROWSER_PROBE_BROWSER_VALUES as readonly string[]).includes(requestedContext.browser)) {
-    throw new Error(`wordpress.browser-probe browser=${requestedContext.browser} is unsupported by this runner; supported browsers: ${BROWSER_PROBE_BROWSER_VALUES.join(", ")}.`)
-  }
-  const deviceProfile = requestedContext.device ? devices[requestedContext.device] : undefined
-  if (requestedContext.device && !deviceProfile) {
-    throw new Error(`wordpress.browser-probe unknown Playwright device profile: ${requestedContext.device}`)
-  }
-  const browser = await launchChromiumBrowser()
-  const browserMetadata = chromiumBrowserMetadata(browser)
-  let finalUrl = targetUrl
-  let windowLocationOrigin: string | undefined
-  let htmlSha256: string | undefined
-  let screenshotSha256: string | undefined
-  let viewport: BrowserProbeViewport | null = null
-  let scriptResult: unknown
-  let lifecycleArtifact: BrowserProbeLifecycleArtifact | undefined
-  let memoryArtifact: BrowserProbeMemoryArtifact | undefined
-  let performanceArtifact: BrowserProbePerformanceArtifact | undefined
-  let page: import("playwright").Page | null = null
-  let context: import("playwright").BrowserContext | null = null
-  let contextDetails: BrowserProbeContextDetails | undefined
-  let authSummary: BrowserProbeAuthSummary | undefined
-  let capabilityDiagnostics: BrowserProbeCapabilityDiagnostics | undefined
-  let assertionResults: import("./browser-artifacts.js").BrowserStepAssertion[] = []
-  let pendingError: Error | undefined
-  let artifact: BrowserProbeArtifact | undefined
-  let wordpressDiagnosticsReady = false
-  const abortHandler = () => {
-    pendingError = pendingError ?? new Error("Browser command aborted during runtime cleanup")
-    void page?.close().catch(() => undefined)
-    void context?.close().catch(() => undefined)
-    void browser.close().catch(() => undefined)
-  }
-  abortSignal?.addEventListener("abort", abortHandler, { once: true })
-
-  try {
-    if (abortSignal?.aborted) {
-      abortHandler()
-      throw pendingError
-    }
-    context = browserPreviewNeedsContextRouting(networkPolicy) || !!storageStateImport || requestedContext.device || requestedContext.locale || requestedContext.timezone || requestedContext.userAgent || (requestedContext.permissions?.length ?? 0) > 0
-      ? await browser.newContext({
-        ...(deviceProfile ?? {}),
-        ...(storageStateImport ? { storageState: storageStateImport.storageState } : {}),
-        ...(requestedContext.locale ? { locale: requestedContext.locale } : {}),
-        ...(requestedContext.timezone ? { timezoneId: requestedContext.timezone } : {}),
-        ...(requestedContext.userAgent ? { userAgent: requestedContext.userAgent } : {}),
-      })
-      : null
-    if (context && requestedContext.permissions && requestedContext.permissions.length > 0) {
-      await context.grantPermissions(requestedContext.permissions)
-    }
-    if (context && browserPreviewNeedsContextRouting(networkPolicy)) {
-      await routeBrowserPreviewContextNetwork(context, networkPolicy, preview.effectiveOrigin, routeTracker)
-    }
-    page = context ? await context.newPage() : await browser.newPage()
-    if (onProgress) {
-      await page.exposeFunction("__wpCodeboxProbeCheckpointEvent", (checkpoint: unknown) => {
-        const normalized = normalizeBrowserProbeScriptCheckpoint(checkpoint)
-        if (!normalized) {
-          return
-        }
-        progress.mark("checkpoint", normalized.timestamp, normalized)
-        onProgress({ command, phase: "checkpoint", checkpoint: normalized, progress: progress.summary() })
-      })
-    }
-    if (storageStateImport) {
-      authSummary = browserStorageStateAuthSummary(storageStateImport.summary)
-    }
-    if (authRequest) {
-      authSummary = await installWordPressAdminAuthCookies({ command, cookieUrls: topology.authCookieUrls([targetUrl]), page, runPlaygroundCommand, runtimeSpec, server, userId: authRequest.userId })
-    }
-    if (requestedViewport) {
-      await page.setViewportSize(requestedViewport)
-    }
-    if (throttleProfile) {
-      await applyBrowserProbeThrottleProfile(page, throttleProfile)
-    }
-    if (!context && browserPreviewNeedsContextRouting(networkPolicy)) {
-      await routeBrowserPreviewPageNetwork(page, networkPolicy, preview.effectiveOrigin, routeTracker)
-    }
-    await page.addInitScript(BROWSER_PROBE_STATE_INIT_SCRIPT)
-    if (lifecycleSelectors.length > 0) {
-      await page.addInitScript(browserProbeLifecycleInitScript(lifecycleSelectors))
-    }
-    if (capturesBrowserMetrics) {
-      await page.addInitScript(BROWSER_PROBE_PERFORMANCE_INIT_SCRIPT)
-    }
-    if (prePageScript) {
-      await page.addInitScript(prePageScript)
-    }
-    wordpressDiagnosticsReady = await installBrowserWordPressDiagnostics(runPlaygroundCommand, server)
-    viewport = await browserProbeViewport(page)
-    contextDetails = await browserProbeContextDetails(page, requestedContext, viewport)
-    capabilityDiagnostics = await browserProbeCapabilityDiagnostics(page, viewport)
-    attachBrowserCaptureListeners({
-      captureConsole: capture.has("console") || capturesConsoleForAssertions,
-      captureErrors: capture.has("errors") || capturesErrorsForAssertions,
-      captureNetwork: true,
-      consoleMessages,
-      errors,
-      network,
-      networkTasks,
-      onConsole: () => progress.mark("console"),
-      onNetwork: () => progress.mark("network"),
-      onPageError: () => progress.mark("pageerror"),
-      page,
-    })
-
-    const previewReadinessError = browserPreviewReadinessError(preview)
-    if (previewReadinessError) {
-      throw previewReadinessError
-    }
-
-    await withBrowserProbeLiveness(page, progress, failFast, navigateBrowserProbe(page, targetUrl, waitFor, durationMs, wallTimeoutMs), livenessPolicy, "navigation")
-    progress.mark("navigation")
-    const browserLocation = await page.evaluate(() => ({ origin: window.location.origin, secureContext: window.isSecureContext })).catch(() => undefined)
-    windowLocationOrigin = browserLocation?.origin
-    preview.secureContext = browserLocation?.secureContext
-    const secureContextError = browserPreviewSecureContextError(preview)
-    if (secureContextError) {
-      throw secureContextError
-    }
-    if (capturesBrowserMetrics) {
-      checkpoints.push(await browserProbeCheckpoint(page, "after-navigation"))
-    }
-    if (script) {
-      scriptResult = await withBrowserProbeLiveness(page, progress, failFast, page.evaluate(async (source) => {
-        const run = new Function(`return (async () => {\n${source}\n})()`)
-        return run()
-      }, script), livenessPolicy, "script")
-      progress.mark("script")
-      if (capturesBrowserMetrics) {
-        const pendingCheckpoints = await browserProbePendingCheckpoints(page)
-        if (pendingCheckpoints.length > 0) {
-          progress.mark("checkpoint")
-        }
-        checkpoints.push(...pendingCheckpoints)
-        checkpoints.push(await browserProbeCheckpoint(page, "after-script"))
-      }
-    }
-    if (durationMs > 0 && waitFor !== "duration") {
-      await withBrowserProbeLiveness(page, progress, failFast, page.waitForTimeout(durationMs), livenessPolicy, "duration")
-      progress.mark("duration")
-      if (capturesBrowserMetrics) {
-        checkpoints.push(await browserProbeCheckpoint(page, "after-duration"))
-      }
-    }
-    if (assertions.length > 0) {
-      await settleBrowserNetworkTasks(networkTasks, livenessPolicy.networkSettleTimeoutMs)
-      const assertionMetrics = capturesBrowserMetrics ? browserProbeBenchMetrics(browserProbeMemoryArtifact(checkpoints), browserProbePerformanceArtifact(checkpoints)) : {}
-      assertionResults = await executeBrowserProbeAssertions(page, assertions, consoleMessages, errors, network, assertionMetrics)
-      if (capturesBrowserMetrics) {
-        checkpoints.push(await browserProbeCheckpoint(page, "after-assertions"))
-      }
-      const fatalFailures = assertionResults.filter((assertion) => !assertion.passed && !assertion.advisory)
-      if (fatalFailures.length > 0) {
-        pendingError = new Error(`wordpress.browser-probe assertion failed: ${fatalFailures.map((assertion) => assertion.assertion).join(", ")}`)
-      }
-    }
-    finalUrl = page.url()
-  } catch (error) {
-    pendingError = error instanceof Error ? error : new Error(String(error))
-    if (isBrowserCommandLivenessError(pendingError)) {
-      await page?.close().catch(() => undefined)
-      page = null
-    }
-    progress.fail("probe-error", pendingError)
-    errors.push(serializeBrowserError("probe-error", error))
-  } finally {
-    if (abortSignal?.aborted) {
-      await closeBrowserBestEffort(browser)
-      abortSignal.removeEventListener("abort", abortHandler)
-      throw pendingError ?? new Error("Browser command aborted during runtime cleanup")
-    }
-    try {
-      await drainBrowserPreviewRouteTracker(routeTracker)
-    } catch (error) {
-      const routeError = error instanceof Error ? error : new Error(String(error))
-      if (!pendingError) {
-        pendingError = routeError
-        progress.fail("probe-error", routeError)
-      }
-      errors.push(serializeBrowserError("probe-error", error))
-    }
-    if (page) {
-      finalUrl = page.url()
-      windowLocationOrigin = windowLocationOrigin ?? await page.evaluate(() => window.location.origin).catch(() => undefined)
-      if (capturesBrowserMetrics) {
-        checkpoints.push(await browserProbeCheckpoint(page, "final"))
-        if (capture.has("memory")) {
-          memoryArtifact = browserProbeMemoryArtifact(checkpoints)
-        }
-        if (capture.has("performance")) {
-          performanceArtifact = browserProbePerformanceArtifact(checkpoints, { consoleMessages, errors, network, startedAt })
-        }
-      }
-      const lifecycle = lifecycleSelectors.length > 0 ? await collectBrowserProbeLifecycle(page) : undefined
-      if (lifecycle) {
-        lifecycleArtifact = browserProbeLifecycleArtifact(lifecycle)
-      }
-
-      if (capture.has("html")) {
-        try {
-          const html = await page.content()
-          await artifactSession.writeText("html", "snapshot.html", html)
-          htmlSha256 = sha256(Buffer.from(html, "utf8"))
-        } catch (error) {
-          errors.push(serializeBrowserError("probe-error", error))
-        }
-      }
-
-      if (capture.has("screenshot")) {
-        try {
-          const activePage = page
-          await artifactSession.writeGenerated("screenshot", "screenshot.png", (path) => activePage.screenshot({ path, fullPage: true }).then(() => undefined))
-          screenshotSha256 = await fileSha256(screenshotPath)
-        } catch (error) {
-          errors.push(serializeBrowserError("probe-error", error))
-        }
-      }
-    }
-    await settleBrowserNetworkTasks(networkTasks, livenessPolicy.networkSettleTimeoutMs)
-    await browser.close()
-    if (capture.has("console") || capturesConsoleForAssertions) {
-      await artifactSession.writeJsonLines("console", "console.jsonl", consoleMessages)
-    }
-    if (capture.has("errors") || capturesErrorsForAssertions) {
-      await artifactSession.writeJsonLines("errors", "errors.jsonl", errors)
-    }
-    if (capture.has("network") || capturesNetworkForAssertions) {
-      await artifactSession.writeJsonLines("network", "network.jsonl", network)
-      await artifactSession.writeJson("waterfall", "waterfall.json", browserProbeWaterfallArtifact(network, startedAt))
-    }
-    if (checkpoints.length > 0) {
-      await artifactSession.writeJsonLines("checkpoints", "checkpoints.jsonl", checkpoints)
-    }
-    if (memoryArtifact) {
-      await artifactSession.writeJson("memory", "memory.json", memoryArtifact)
-    }
-    if (lifecycleArtifact) {
-      await artifactSession.writeJson("lifecycle", "lifecycle.json", lifecycleArtifact)
-    }
-    if (performanceArtifact) {
-      await artifactSession.writeJson("performance", "performance.json", performanceArtifact)
-    }
-
-    const redirectDiagnostics = browserRedirectDiagnosticsArtifact({
-      artifactPath: `${browserFilesDirectory}/redirect-diagnostics.json`,
-      error: pendingError,
-      finalAttemptedUrl: finalUrl,
-      network,
-      requestedUrl: targetUrl,
-    })
-    if (redirectDiagnostics) {
-      await artifactSession.writeJson("redirectDiagnostics", "redirect-diagnostics.json", redirectDiagnostics)
-    }
-    const redirectDiagnosticsSummary = redirectDiagnostics?.summary
-
-    const wordpressDiagnostics = await browserWordPressDiagnosticsArtifact({
-      artifactPath: `${browserFilesDirectory}/wordpress-diagnostics.json`,
-      network,
-      ready: wordpressDiagnosticsReady,
-      server,
-    })
-    if (wordpressDiagnostics) {
-      await artifactSession.writeJson("wordpressDiagnostics", "wordpress-diagnostics.json", wordpressDiagnostics)
-    }
-    const wordpressDiagnosticsSummary = wordpressDiagnostics?.summary
-
-    const assertionPassed = assertionResults.filter((assertion) => assertion.passed).length
-    const assertionFailed = assertionResults.filter((assertion) => !assertion.passed).length
-    const advisoryFailed = assertionResults.filter((assertion) => !assertion.passed && assertion.advisory).length
-    const assertionSummary = {
-      total: assertionResults.length,
-      passed: assertionPassed,
-      failed: assertionFailed,
-      advisoryFailed,
-      fatalFailed: assertionFailed - advisoryFailed,
-      results: assertionResults,
-    }
-
-    const finishedAt = now()
-    const review = browserProbeReviewSummary({
-      browser: browserMetadata,
-      capture,
-      checkpoints,
-      consoleMessages,
-      durationMs,
-      errors,
-      files: browserProbeArtifactRefs(browserFilesDirectory, capture, {
-        checkpoints: checkpoints.length > 0,
-        console: capture.has("console") || capturesConsoleForAssertions,
-        errors: capture.has("errors") || capturesErrorsForAssertions,
-        html: capture.has("html") ? htmlSha256 : undefined,
-        lifecycle: Boolean(lifecycleArtifact),
-        memory: Boolean(memoryArtifact),
-        network: capture.has("network") || capturesNetworkForAssertions,
-        waterfall: capture.has("network") || capturesNetworkForAssertions,
-        performance: Boolean(performanceArtifact),
-        redirectDiagnostics: Boolean(redirectDiagnostics),
-        screenshot: capture.has("screenshot") ? screenshotSha256 : undefined,
-        wordpressDiagnostics: Boolean(wordpressDiagnostics),
-      }),
-      finishedAt,
-      network,
-      performanceArtifact,
-      startedAt,
-      throttle: throttleProfile?.id ?? null,
-      totalDurationMs: Date.now() - startedAtMs,
-      viewport,
-      waitFor,
-      redirectDiagnostics: redirectDiagnosticsSummary,
-      wordpressDiagnostics: wordpressDiagnosticsSummary,
-    })
-    await artifactSession.writeJson("review", "review.json", review)
-
-    artifact = {
-      artifactType: "probe",
-      requestedUrl: targetUrl,
-      url: targetUrl,
-      preview,
-      ...(server.previewProxyDiagnostics ? { previewProxy: server.previewProxyDiagnostics } : {}),
-      ...(browserPreviewNetworkPolicyIsActive(networkPolicy) ? { networkPolicy: browserPreviewNetworkPolicySummary(networkPolicy) } : {}),
-      ...topology.origins,
-      ...(prePageScriptMetadata ? { prePageScript: prePageScriptMetadata } : {}),
-      files: {
-        ...(capture.has("console") || capturesConsoleForAssertions ? { console: `${browserFilesDirectory}/console.jsonl` } : {}),
-        ...(checkpoints.length > 0 ? { checkpoints: `${browserFilesDirectory}/checkpoints.jsonl` } : {}),
-        ...(capture.has("errors") || capturesErrorsForAssertions ? { errors: `${browserFilesDirectory}/errors.jsonl` } : {}),
-        ...(htmlSha256 ? { html: `${browserFilesDirectory}/snapshot.html` } : {}),
-        ...(lifecycleArtifact ? { lifecycle: `${browserFilesDirectory}/lifecycle.json` } : {}),
-        ...(memoryArtifact ? { memory: `${browserFilesDirectory}/memory.json` } : {}),
-        ...(capture.has("network") || capturesNetworkForAssertions ? { network: `${browserFilesDirectory}/network.jsonl` } : {}),
-        ...(capture.has("network") || capturesNetworkForAssertions ? { waterfall: `${browserFilesDirectory}/waterfall.json` } : {}),
-        ...(performanceArtifact ? { performance: `${browserFilesDirectory}/performance.json` } : {}),
-        ...(redirectDiagnostics ? { redirectDiagnostics: `${browserFilesDirectory}/redirect-diagnostics.json` } : {}),
-        review: `${browserFilesDirectory}/review.json`,
-        ...(screenshotSha256 ? { screenshot: `${browserFilesDirectory}/screenshot.png` } : {}),
-        ...(wordpressDiagnostics ? { wordpressDiagnostics: `${browserFilesDirectory}/wordpress-diagnostics.json` } : {}),
-        summary: `${browserFilesDirectory}/summary.json`,
-      },
-      summary: {
-        ...(assertionSummary.total > 0 ? { assertions: assertionSummary } : {}),
-        consoleMessages: consoleMessages.length,
-        errors: errors.length,
-        finalUrl,
-        ...(windowLocationOrigin ? { windowLocationOrigin } : {}),
-        htmlSnapshot: Boolean(htmlSha256),
-        ...(server.previewProxyDiagnostics ? { previewProxy: server.previewProxyDiagnostics } : {}),
-        ...(browserPreviewNetworkPolicyIsActive(networkPolicy) ? { networkPolicy: browserPreviewNetworkPolicySummary(networkPolicy) } : {}),
-        ...(lifecycleArtifact ? { lifecycle: { schema: lifecycleArtifact.schema, version: lifecycleArtifact.version, startedAtMs: lifecycleArtifact.startedAtMs, selectors: lifecycleArtifact.selectors } } : {}),
-        liveness: { wallTimeoutMs, stallTimeoutMs, networkSettleTimeoutMs: livenessPolicy.networkSettleTimeoutMs },
-        ...(memoryArtifact ? { memory: memoryArtifact.peak } : {}),
-        ...(memoryArtifact || performanceArtifact ? { metrics: browserProbeBenchMetrics(memoryArtifact, performanceArtifact) } : {}),
-        networkEvents: network.length,
-        ...(performanceArtifact?.phaseMetrics ? { phaseMetrics: performanceArtifact.phaseMetrics } : {}),
-        ...(performanceArtifact ? { performance: performanceArtifact.peak } : {}),
-        progress: progress.summary(),
-        review,
-        ...(redirectDiagnosticsSummary ? { redirectDiagnostics: redirectDiagnosticsSummary } : {}),
-        ...(wordpressDiagnosticsSummary ? { wordpressDiagnostics: wordpressDiagnosticsSummary } : {}),
-        context: contextDetails,
-        auth: authSummary,
-        capabilities: capabilityDiagnostics,
-        replayability: browserProbeReplayability(capture),
-        screenshot: Boolean(screenshotSha256),
-        ...(typeof scriptResult !== "undefined" ? { scriptResult } : {}),
-        viewport,
-      },
-    }
-    await artifactSession.writeJson("summary", "summary.json", {
-      schema: "wp-codebox/browser-probe/v1",
-      requestedUrl: targetUrl,
-      preview,
-      ...(server.previewProxyDiagnostics ? { previewProxy: server.previewProxyDiagnostics } : {}),
-      ...(browserPreviewNetworkPolicyIsActive(networkPolicy) ? { networkPolicy: browserPreviewNetworkPolicySummary(networkPolicy) } : {}),
-      ...topology.origins,
-      finalUrl,
-      ...(windowLocationOrigin ? { windowLocationOrigin } : {}),
-      waitFor,
-      durationMs,
-      ...(lifecycleSelectors.length > 0 ? { observe: lifecycleSelectors } : {}),
-      failFast,
-      stallTimeoutMs,
-      capture: [...capture].sort(),
-      ...(assertionSummary.total > 0 ? { assertions: assertionSummary } : {}),
-      ...(prePageScriptMetadata ? { prePageScript: prePageScriptMetadata } : {}),
-      startedAt,
-      finishedAt,
-      files: artifact.files,
-      hashes: {
-        ...(htmlSha256 ? { html: { algorithm: "sha256", value: htmlSha256 } } : {}),
-        ...(screenshotSha256 ? { screenshot: { algorithm: "sha256", value: screenshotSha256 } } : {}),
-      },
-      context: contextDetails,
-      auth: authSummary,
-      capabilities: capabilityDiagnostics,
-      review,
-      ...(redirectDiagnosticsSummary ? { redirectDiagnostics: redirectDiagnosticsSummary } : {}),
-      ...(wordpressDiagnosticsSummary ? { wordpressDiagnostics: wordpressDiagnosticsSummary } : {}),
-      viewport,
-      summary: artifact.summary,
-    })
-  }
-
-  abortSignal?.removeEventListener("abort", abortHandler)
-  if (pendingError) {
-    if (!artifact) {
-      throw pendingError
-    }
-    throw new BrowserCommandArtifactError(pendingError.message, artifact)
-  }
-  if (!artifact) {
-    throw new Error("wordpress.browser-probe did not produce a browser artifact")
-  }
-  return {
-    artifact,
-    output: `${JSON.stringify({
-      command,
-      requestedUrl: targetUrl,
-      preview,
-      ...(server.previewProxyDiagnostics ? { previewProxy: server.previewProxyDiagnostics } : {}),
-      ...(browserPreviewNetworkPolicyIsActive(networkPolicy) ? { networkPolicy: browserPreviewNetworkPolicySummary(networkPolicy) } : {}),
-      ...topology.origins,
-      finalUrl: artifact.summary.finalUrl ?? targetUrl,
-      files: artifact.files,
-      summary: artifact.summary,
-    }, null, 2)}\n`,
-  }
-}
-
-async function closeBrowserBestEffort(browser: import("playwright").Browser): Promise<void> {
-  await Promise.race([
-    browser.close().catch(() => undefined),
-    new Promise<void>((resolve) => {
-      const timeout = setTimeout(resolve, 1_000)
-      timeout.unref()
-    }),
-  ])
-}
-
-function browserProbeProfileIds(args: string[]): string[] {
-  const raw = argValue(args, "profiles")?.trim()
-  if (!raw) {
-    return []
-  }
-  return raw.split(",").map((profile) => profile.trim()).filter(Boolean)
-}
-
-function browserProbeProfile(profileId: string): BrowserProbeProfileDefinition {
-  const profile = BROWSER_PROBE_PROFILES[profileId as keyof typeof BROWSER_PROBE_PROFILES]
-  if (!profile) {
-    throw new Error(`wordpress.browser-probe unknown profile: ${profileId}. Supported profiles: ${BROWSER_PROBE_CHROMIUM_PROFILE_IDS.join(", ")}`)
-  }
-  return profile
-}
-
-async function browserProbeRunPlanFromArgs(args: string[], profileId?: string): Promise<BrowserProbeRunPlan> {
-  const capture = new Set(commaListArg(args, "capture"))
-  if (capture.size === 0) {
-    capture.add("console")
-    capture.add("errors")
-    capture.add("html")
-    capture.add("network")
-    capture.add("screenshot")
-  }
-  const requestedViewport = viewportArg(args, "viewport")
-  const throttleProfile = browserProbeThrottleProfile(args)
-  return {
-    url: argValue(args, "url")?.trim() ?? "",
-    capture,
-    waitFor: argValue(args, "wait-for")?.trim() || "domcontentloaded",
-    durationMs: durationArg(args, "duration", 0),
-    requestedViewport,
-    throttleProfile,
-    requestedContext: browserProbeContextRequest(args, requestedViewport, profileId, throttleProfile?.id),
-    prePageScript: argValue(args, "pre-page-script"),
-    script: argValue(args, "script"),
-    authRequest: browserAuthRequest(args),
-    storageStateImport: await browserStorageStateImportFromArgs(args, "wordpress.browser-probe"),
-    failFast: strictBooleanArg(args, "fail-fast", false),
-    stallTimeoutMs: durationArg(args, "stall-timeout", 0),
-    wallTimeoutMs: durationArg(args, "timeout", browserCommandLivenessPolicy().wallTimeoutMs),
-    lifecycleSelectors: commaListArg(args, "observe"),
-    assertions: browserProbeAssertionsFromArgs(args),
-  }
-}
-
-function browserProbeProfileArgs(args: string[], profile: BrowserProbeProfileDefinition): string[] {
-  const explicitOverrideKeys = new Set(args.map((arg) => arg.match(/^([^=]+)=/)?.[1]).filter((key): key is string => typeof key === "string" && BROWSER_PROBE_PROFILE_OVERRIDES.has(key)))
-  return [
-    ...args.filter((arg) => !arg.startsWith("profiles=") && !arg.startsWith("profile=")),
-    `profile=${profile.id}`,
-    ...profile.args.filter((arg) => {
-      const key = arg.match(/^([^=]+)=/)?.[1]
-      return !key || !explicitOverrideKeys.has(key)
-    }),
-  ]
-}
-
-function browserProbeThrottleProfile(args: string[]): BrowserProbeThrottleProfileDefinition | undefined {
-  const profileId = argValue(args, "throttle")?.trim()
-  if (!profileId || profileId === "none") {
-    return undefined
-  }
-
-  const profile = BROWSER_PROBE_THROTTLE_PROFILES[profileId]
-  if (!profile) {
-    throw new Error(`wordpress.browser-probe unknown throttle profile: ${profileId}. Supported profiles: ${BROWSER_PROBE_THROTTLE_PROFILE_IDS.join(", ")}`)
-  }
-  return profile
-}
-
-async function applyBrowserProbeThrottleProfile(page: import("playwright").Page, profile: BrowserProbeThrottleProfileDefinition): Promise<void> {
-  const session = await page.context().newCDPSession(page)
-  try {
-    await Promise.all([
-      session.send("Network.enable").catch(() => undefined),
-      session.send("Emulation.setCPUThrottlingRate", { rate: profile.cpuSlowdownRate }).catch(() => undefined),
-    ])
-    await session.send("Network.emulateNetworkConditions", {
-      offline: profile.network.offline,
-      latency: profile.network.latencyMs,
-      downloadThroughput: profile.network.downloadThroughputBytesPerSecond,
-      uploadThroughput: profile.network.uploadThroughputBytesPerSecond,
-    }).catch(() => undefined)
-  } finally {
-    await session.detach().catch(() => undefined)
-  }
-}
-
-function browserProbeScriptMetadata(source: string): BrowserProbeScriptMetadata {
-  return {
-    sha256: sha256(Buffer.from(source, "utf8")),
-    bytes: Buffer.byteLength(source, "utf8"),
-  }
-}
-
-function browserProbeContextRequest(args: string[], viewport: { width: number; height: number } | undefined, profileId?: string, throttleProfileId?: string): BrowserProbeContextDetails["requested"] {
-  const browser = argValue(args, "browser")?.trim()
-  const device = argValue(args, "device")?.trim()
-  const locale = argValue(args, "locale")?.trim()
-  const permissions = commaListArg(args, "permissions")
-  const timezone = argValue(args, "timezone")?.trim()
-  const userAgent = argValue(args, "user-agent")?.trim()
-  return {
-    ...(browser ? { browser } : {}),
-    ...(device ? { device } : {}),
-    ...(locale ? { locale } : {}),
-    ...(permissions.length > 0 ? { permissions } : {}),
-    ...(profileId ? { profile: profileId } : {}),
-    ...(throttleProfileId ? { throttle: throttleProfileId } : {}),
-    ...(timezone ? { timezone } : {}),
-    ...(userAgent ? { userAgent } : {}),
-    ...(viewport ? { viewport } : {}),
-  }
-}
-
-async function browserProbeContextDetails(page: import("playwright").Page, requested: BrowserProbeContextDetails["requested"], viewport: BrowserProbeViewport | null): Promise<BrowserProbeContextDetails> {
-  const effective = await page.evaluate(() => ({
-    locale: navigator.language || undefined,
-    timezone: Intl.DateTimeFormat().resolvedOptions().timeZone || undefined,
-  })).catch(() => ({ locale: undefined, timezone: undefined }))
-
-  return {
-    requested,
-    effective: {
-      ...(requested.browser ? { browser: requested.browser } : {}),
-      ...(requested.device ? { device: requested.device } : {}),
-      ...(effective.locale ? { locale: effective.locale } : {}),
-      ...(requested.permissions ? { permissions: requested.permissions } : {}),
-      ...(requested.profile ? { profile: requested.profile } : {}),
-      ...(requested.throttle ? { throttle: requested.throttle } : {}),
-      ...(effective.timezone ? { timezone: effective.timezone } : {}),
-      ...(viewport?.userAgent ? { userAgent: viewport.userAgent } : {}),
-      viewport,
-    },
-  }
-}
-
-async function browserProbeCapabilityDiagnostics(page: import("playwright").Page, viewport: BrowserProbeViewport | null): Promise<BrowserProbeCapabilityDiagnostics> {
-  const fallback: BrowserProbeCapabilityDiagnostics = {
-    secureContext: false,
-    userAgent: viewport?.userAgent ?? "",
-    viewport: viewport ? browserProbeCapabilityViewport(viewport) : null,
-    maxTouchPoints: 0,
-    paymentRequest: { available: false },
-    permissions: {},
-  }
-
-  return page.evaluate(async (probeViewport) => {
-    const resolvedOptions = Intl.DateTimeFormat().resolvedOptions()
-    const permissionNames = ["clipboard-read", "clipboard-write", "geolocation", "notifications", "payment-handler"]
-    const permissions: BrowserProbeCapabilityDiagnostics["permissions"] = {}
-    const permissionsApi = (navigator as typeof navigator & { permissions?: { query?: (descriptor: { name: string }) => Promise<{ state?: string }> } }).permissions
-    const normalizePermissionState = (state: unknown): BrowserProbeCapabilityDiagnostics["permissions"][string]["state"] => {
-      if (state === "granted" || state === "denied" || state === "prompt") {
-        return state
-      }
-      return typeof state === "string" ? "error" : "unsupported"
-    }
-
-    for (const name of permissionNames) {
-      if (!permissionsApi?.query) {
-        permissions[name] = { state: "unsupported" }
-        continue
-      }
-
-      try {
-        const status = await permissionsApi.query({ name })
-        permissions[name] = { state: normalizePermissionState(status.state) }
-      } catch {
-        permissions[name] = { state: "unsupported" }
-      }
-    }
-
-    return {
-      secureContext: Boolean(window.isSecureContext),
-      userAgent: navigator.userAgent || "",
-      language: navigator.language || undefined,
-      languages: Array.isArray(navigator.languages) ? navigator.languages.slice(0, 10).filter((language) => typeof language === "string" && language.length > 0) : undefined,
-      locale: resolvedOptions.locale || navigator.language || undefined,
-      timezone: resolvedOptions.timeZone || undefined,
-      viewport: probeViewport ? {
-        width: probeViewport.width,
-        height: probeViewport.height,
-        deviceScaleFactor: probeViewport.deviceScaleFactor,
-        isMobile: probeViewport.isMobile,
-        hasTouch: probeViewport.hasTouch,
-      } : null,
-      maxTouchPoints: typeof navigator.maxTouchPoints === "number" ? navigator.maxTouchPoints : 0,
-      paymentRequest: { available: typeof (window as typeof window & { PaymentRequest?: unknown }).PaymentRequest === "function" },
-      permissions,
-    }
-  }, viewport).catch(() => fallback)
-}
-
-function browserProbeCapabilityViewport(viewport: BrowserProbeViewport): BrowserProbeCapabilityDiagnostics["viewport"] {
-  return {
-    width: viewport.width,
-    height: viewport.height,
-    deviceScaleFactor: viewport.deviceScaleFactor,
-    isMobile: viewport.isMobile,
-    hasTouch: viewport.hasTouch,
-  }
-}
-
-function browserProbeReviewSummary(input: {
-  browser: BrowserProbeReviewSummary["browser"]
-  capture: Set<string>
-  checkpoints: BrowserProbeCheckpointRecord[]
-  consoleMessages: Record<string, unknown>[]
-  durationMs: number
-  errors: BrowserProbeErrorRecord[]
-  files: Record<string, BrowserProbeArtifactRef>
-  finishedAt: string
-  network: BrowserProbeNetworkRecord[]
-  performanceArtifact?: BrowserProbePerformanceArtifact
-  startedAt: string
-  throttle: string | null
-  totalDurationMs: number
-  viewport: BrowserProbeViewport | null
-  waitFor: string
-  redirectDiagnostics?: BrowserRedirectDiagnosticsSummary
-  wordpressDiagnostics?: BrowserWordPressDiagnosticsSummary
-}): BrowserProbeReviewSummary {
-  const performance = input.performanceArtifact?.final
-  const paint = performance?.paint
-  const navigation = performance?.navigation ?? {
-    type: null,
-    redirectCount: 0,
-    durationMs: null,
-    domContentLoadedMs: null,
-    loadEventMs: null,
-    responseStartMs: null,
-    responseEndMs: null,
-    requestStartMs: null,
-    ttfbMs: null,
-    redirectMs: null,
-  }
-  const missingReason = input.capture.has("performance")
-    ? "metric not reported by browser"
-    : "capture=performance was not requested"
-  const pageErrors = input.errors.filter((error) => error.type === "pageerror")
-  const probeErrors = input.errors.filter((error) => error.type === "probe-error")
-  const consoleErrors = input.consoleMessages.filter((message) => message.type === "error")
-  const lcpUrl = safeBrowserProbeUrl(paint?.largestContentfulPaintUrl ?? undefined)
-
-  return {
-    schema: "wp-codebox/browser-probe-review/v1",
-    version: 1,
-    browser: input.browser,
-    runtime: {
-      node: process.version,
-      platform: process.platform,
-      arch: process.arch,
-    },
-    profile: {
-      viewport: input.viewport,
-      userAgent: input.viewport?.userAgent ?? null,
-      throttle: input.throttle,
-      waitFor: input.waitFor,
-      durationMs: input.durationMs,
-    },
-    timings: {
-      startedAt: input.startedAt,
-      finishedAt: input.finishedAt,
-      totalDurationMs: browserProbeMetric(input.totalDurationMs, "probe wall-clock duration unavailable"),
-      navigation,
-      ttfbMs: browserProbeMetric(navigation.ttfbMs, missingReason),
-      firstContentfulPaintMs: browserProbeMetric(paint?.firstContentfulPaintMs, missingReason),
-      largestContentfulPaintMs: browserProbeMetric(paint?.largestContentfulPaintMs, missingReason),
-      loadEventMs: browserProbeMetric(navigation.loadEventMs, missingReason),
-    },
-    lcp: {
-      status: typeof paint?.largestContentfulPaintMs === "number" ? "available" : "missing",
-      ...(typeof paint?.largestContentfulPaintMs === "number" ? {} : { reason: missingReason }),
-      element: paint?.largestContentfulPaintElement ?? null,
-      size: paint?.largestContentfulPaintSize ?? null,
-      url: lcpUrl ?? null,
-    },
-    errors: {
-      console: browserProbeIssueSummary(consoleErrors.length, Boolean(input.files.console), input.files.console?.path),
-      page: browserProbeIssueSummary(pageErrors.length, Boolean(input.files.errors), input.files.errors?.path),
-      probe: browserProbeIssueSummary(probeErrors.length, Boolean(input.files.errors), input.files.errors?.path),
-    },
-    network: browserProbeNetworkReviewSummary(input.network, input.files.network, input.files.waterfall),
-    ...(input.redirectDiagnostics ? { redirectDiagnostics: input.redirectDiagnostics } : {}),
-    ...(input.wordpressDiagnostics ? { wordpressDiagnostics: input.wordpressDiagnostics } : {}),
-    milestones: {
-      status: input.checkpoints.length > 0 ? "captured" : "not-captured",
-      count: input.checkpoints.length,
-      names: [...new Set(input.checkpoints.map((checkpoint) => checkpoint.name).filter(Boolean))].sort(),
-      ...(input.files.checkpoints ? { artifact: input.files.checkpoints.path } : {}),
-    },
-    artifacts: input.files,
-  }
-}
-
-function browserProbeMetric(value: number | null | undefined, reason: string): BrowserProbeMeasuredMetric {
-  return typeof value === "number" && Number.isFinite(value)
-    ? { status: "available", value, unit: "ms" }
-    : { status: "missing", value: null, unit: "ms", reason }
-}
-
-function browserProbeIssueSummary(count: number, captured: boolean, artifact?: string): BrowserProbeReviewSummary["errors"]["console"] {
-  if (!captured && count === 0) {
-    return { count: 0, status: "not-captured" }
-  }
-  return {
-    count,
-    status: count > 0 ? "issues" : "ok",
-    ...(artifact ? { artifact } : {}),
-  }
-}
-
-function browserProbeNetworkReviewSummary(network: BrowserProbeNetworkRecord[], networkArtifact?: BrowserProbeArtifactRef, waterfallArtifact?: BrowserProbeArtifactRef): BrowserProbeNetworkReviewSummary {
-  if (!networkArtifact) {
-    return { status: "not-captured", events: 0, responses: 0, failures: 0, byHost: {}, byType: {} }
-  }
-
-  const byHost: Record<string, BrowserProbeNetworkCountSummary> = {}
-  const byType: Record<string, BrowserProbeNetworkCountSummary> = {}
-  for (const record of network) {
-    addBrowserProbeNetworkCount(byHost, requestHost(record.url) || "unknown", record)
-    addBrowserProbeNetworkCount(byType, record.resourceType || "unknown", record)
-  }
-
-  return {
-    status: "captured",
-    events: network.length,
-    responses: network.filter((record) => record.type === "response").length,
-    failures: network.filter((record) => record.type === "requestfailed").length,
-    byHost: sortBrowserProbeNetworkCounts(byHost),
-    byType: sortBrowserProbeNetworkCounts(byType),
-    ...(waterfallArtifact ? { waterfall: waterfallArtifact } : {}),
-  }
-}
-
-function browserProbeWaterfallArtifact(network: BrowserProbeNetworkRecord[], startedAt: string): BrowserProbeWaterfallArtifact {
-  return {
-    schema: "wp-codebox/browser-waterfall/v1",
-    version: 1,
-    capturedAt: now(),
-    startedAt,
-    summary: {
-      requests: network.length,
-      responses: network.filter((record) => record.type === "response").length,
-      failures: network.filter((record) => record.type === "requestfailed").length,
-      transferSizeBytes: network.reduce((total, record) => total + finiteNumber(record.transferSize, 0), 0),
-    },
-    log: {
-      version: "1.2",
-      creator: { name: "wp-codebox", version: "1" },
-      entries: network.map(browserProbeWaterfallEntry),
-    },
-  }
-}
-
-function browserProbeWaterfallEntry(record: BrowserProbeNetworkRecord): BrowserProbeWaterfallEntry {
-  const timings = browserProbeWaterfallTimings(record.timing ?? {})
-  const startedDateTime = browserProbeWaterfallStartedDateTime(record)
-  const responseEnd = finiteNumber(record.timing?.responseEnd, 0)
-  const fallbackTime = Math.max(0, Date.parse(record.timestamp) - Date.parse(startedDateTime))
-  const time = responseEnd > 0 ? responseEnd : fallbackTime
-  return {
-    startedDateTime,
-    time,
-    request: {
-      method: record.method,
-      url: redactBrowserArtifactUrl(record.url),
-    },
-    response: {
-      status: record.status ?? 0,
-      statusText: record.statusText ?? (record.type === "requestfailed" ? "Request Failed" : ""),
-      content: { mimeType: record.contentType ?? "" },
-      redirectURL: "",
-    },
-    cache: {},
-    timings,
-    _wpCodebox: {
-      type: record.type,
-      resourceType: record.resourceType,
-      timestamp: record.timestamp,
-      ...(typeof record.ok === "boolean" ? { ok: record.ok } : {}),
-      ...(typeof record.transferSize === "number" ? { transferSize: record.transferSize } : {}),
-      ...(typeof record.requestBodySize === "number" ? { requestBodySize: record.requestBodySize } : {}),
-      ...(typeof record.responseBodySize === "number" ? { responseBodySize: record.responseBodySize } : {}),
-      ...(record.failure ? { failure: record.failure } : {}),
-    },
-  }
-}
-
-function browserProbeWaterfallTimings(timing: Record<string, number>): BrowserProbeWaterfallEntry["timings"] {
-  const requestStart = finiteNumber(timing.requestStart, 0)
-  const responseStart = finiteNumber(timing.responseStart, 0)
-  const responseEnd = finiteNumber(timing.responseEnd, responseStart)
-  const dns = timingDelta(timing.domainLookupStart, timing.domainLookupEnd)
-  const connect = timingDelta(timing.connectStart, timing.connectEnd)
-  const ssl = timingDelta(timing.secureConnectionStart, timing.connectEnd)
-  return {
-    blocked: Math.max(0, requestStart),
-    dns,
-    connect,
-    ssl,
-    send: timingDelta(timing.requestStart, timing.requestStart),
-    wait: responseStart >= requestStart ? responseStart - requestStart : 0,
-    receive: responseEnd >= responseStart ? responseEnd - responseStart : 0,
-  }
-}
-
-function browserProbeWaterfallStartedDateTime(record: BrowserProbeNetworkRecord): string {
-  const startTime = record.timing?.startTime
-  if (typeof startTime === "number" && Number.isFinite(startTime) && startTime > 0) {
-    return new Date(startTime).toISOString()
-  }
-  return record.timestamp
-}
-
-function timingDelta(start: number | undefined, end: number | undefined): number {
-  if (typeof start !== "number" || typeof end !== "number" || !Number.isFinite(start) || !Number.isFinite(end) || start < 0 || end < start) {
-    return -1
-  }
-  return end - start
-}
-
-function finiteNumber(value: number | undefined, fallback: number): number {
-  return typeof value === "number" && Number.isFinite(value) ? value : fallback
-}
-
-function redactBrowserArtifactUrl(url: string): string {
-  return redactString(url, { redactAllUrlQueryValues: true, redactUrlHash: true, redactQueryAssignments: true })
-}
-
-function addBrowserProbeNetworkCount(target: Record<string, BrowserProbeNetworkCountSummary>, key: string, record: BrowserProbeNetworkRecord): void {
-  const summary = target[key] ?? { requests: 0, responses: 0, failures: 0, transferSizeBytes: 0 }
-  summary.requests += 1
-  if (record.type === "response") {
-    summary.responses += 1
-  }
-  if (record.type === "requestfailed") {
-    summary.failures += 1
-  }
-  summary.transferSizeBytes += typeof record.transferSize === "number" && Number.isFinite(record.transferSize) ? record.transferSize : 0
-  target[key] = summary
-}
-
-function sortBrowserProbeNetworkCounts(value: Record<string, BrowserProbeNetworkCountSummary>): Record<string, BrowserProbeNetworkCountSummary> {
-  return Object.fromEntries(Object.entries(value).sort(([left], [right]) => left.localeCompare(right)))
-}
-
-function requestHost(url: string): string | undefined {
-  try {
-    return new URL(url).host
-  } catch {
-    return undefined
-  }
-}
-
-function browserProbeArtifactRefs(browserFilesDirectory: string, capture: Set<string>, input: {
-  checkpoints: boolean
-  console: boolean
-  errors: boolean
-  html?: string
-  lifecycle: boolean
-  memory: boolean
-  network: boolean
-  waterfall: boolean
-  performance: boolean
-  redirectDiagnostics: boolean
-  screenshot?: string
-  wordpressDiagnostics: boolean
-}): Record<string, BrowserProbeArtifactRef> {
-  return {
-    ...(input.console ? { console: { path: `${browserFilesDirectory}/console.jsonl`, kind: "jsonl" as const } } : {}),
-    ...(input.checkpoints ? { checkpoints: { path: `${browserFilesDirectory}/checkpoints.jsonl`, kind: "jsonl" as const } } : {}),
-    ...(input.errors ? { errors: { path: `${browserFilesDirectory}/errors.jsonl`, kind: "jsonl" as const } } : {}),
-    ...(input.html ? { html: { path: `${browserFilesDirectory}/snapshot.html`, kind: "html" as const, sha256: input.html } } : {}),
-    ...(input.lifecycle ? { lifecycle: { path: `${browserFilesDirectory}/lifecycle.json`, kind: "json" as const } } : {}),
-    ...(input.memory ? { memory: { path: `${browserFilesDirectory}/memory.json`, kind: "json" as const } } : {}),
-    ...(input.network ? { network: { path: `${browserFilesDirectory}/network.jsonl`, kind: "jsonl" as const } } : {}),
-    ...(input.waterfall ? { waterfall: { path: `${browserFilesDirectory}/waterfall.json`, kind: "json" as const } } : {}),
-    ...(input.performance ? { performance: { path: `${browserFilesDirectory}/performance.json`, kind: "json" as const } } : {}),
-    ...(input.redirectDiagnostics ? { redirectDiagnostics: { path: `${browserFilesDirectory}/redirect-diagnostics.json`, kind: "json" as const } } : {}),
-    review: { path: `${browserFilesDirectory}/review.json`, kind: "json" as const },
-    ...(capture.has("screenshot") ? { screenshot: { path: `${browserFilesDirectory}/screenshot.png`, kind: "png" as const, ...(input.screenshot ? { sha256: input.screenshot } : {}) } } : {}),
-    ...(input.wordpressDiagnostics ? { wordpressDiagnostics: { path: `${browserFilesDirectory}/wordpress-diagnostics.json`, kind: "json" as const } } : {}),
-    summary: { path: `${browserFilesDirectory}/summary.json`, kind: "json" as const },
-  }
-}
-
-function safeBrowserProbeUrl(value: string | undefined): string | null {
-  if (!value) {
-    return null
-  }
-  if (/^data:/i.test(value)) {
-    return "data:[redacted]"
-  }
-  return value
-}
-
-interface BrowserRedirectDiagnosticsArtifact {
-  schema: "wp-codebox/browser-redirect-diagnostics/v1"
-  version: 1
-  capturedAt: string
-  status: BrowserRedirectDiagnosticsSummary["status"]
-  classification: BrowserRedirectDiagnosticsSummary["classification"]
-  reason: string
-  error?: { name: string; message: string }
-  chain: BrowserRedirectDiagnosticsChainEntry[]
-  summary: BrowserRedirectDiagnosticsSummary
-}
-
-interface BrowserRedirectDiagnosticsChainEntry {
-  url: string
-  method: string
-  status?: number
-  statusText?: string
-  timestamp: string
-  host?: string
-  path?: string
-  queryKeys: string[]
-  redactedQueryKeys: string[]
-}
-
-function browserRedirectDiagnosticsArtifact({
-  artifactPath,
-  error,
-  finalAttemptedUrl,
-  network,
-  requestedUrl,
-}: {
-  artifactPath: string
-  error?: Error
-  finalAttemptedUrl: string
-  network: BrowserProbeNetworkRecord[]
-  requestedUrl: string
-}): BrowserRedirectDiagnosticsArtifact | undefined {
-  const errorMessage = error?.message ?? ""
-  const tooManyRedirects = /ERR_TOO_MANY_REDIRECTS/i.test(errorMessage)
-  const documentEvents = network.filter((record) => record.resourceType === "document")
-  const redirectResponses = documentEvents.filter((record) => record.type === "response" && typeof record.status === "number" && record.status >= 300 && record.status < 400)
-  const chain = documentEvents.map(browserRedirectDiagnosticsChainEntry)
-  const repeatedUrls = repeatedBrowserRedirectValues(chain.map((entry) => entry.url), "url")
-  const repeatedHosts = repeatedBrowserRedirectValues(chain.map((entry) => entry.host).filter((host): host is string => Boolean(host)), "host")
-  const repeatedPaths = repeatedBrowserRedirectValues(chain.map((entry) => entry.path).filter((path): path is string => Boolean(path)), "path")
-  const hasRepeatedTarget = repeatedUrls.length > 0 || repeatedHosts.length > 0 || repeatedPaths.length > 0
-
-  if (!tooManyRedirects && redirectResponses.length === 0 && !hasRepeatedTarget) {
-    return undefined
-  }
-
-  const finalAttempted = browserRedirectSafeUrl(extractBrowserNavigationUrl(errorMessage) ?? finalAttemptedUrl)
-  const firstUrl = chain[0]?.url ?? browserRedirectSafeUrl(requestedUrl)
-  const lastUrl = chain.at(-1)?.url ?? finalAttempted
-  const sanitizedQueryKeys = [...new Set(chain.flatMap((entry) => entry.queryKeys))].sort()
-  const redactedQueryKeys = [...new Set(chain.flatMap((entry) => entry.redactedQueryKeys))].sort()
-  const classification: BrowserRedirectDiagnosticsSummary["classification"] = tooManyRedirects || hasRepeatedTarget ? "redirect-loop" : "redirect-chain"
-  const reason = tooManyRedirects
-    ? "playwright reported ERR_TOO_MANY_REDIRECTS"
-    : hasRepeatedTarget ? "document navigation repeated URL, host, or path values" : "document navigation included redirect responses"
-  const summary: BrowserRedirectDiagnosticsSummary = {
-    status: "captured",
-    artifact: artifactPath,
-    classification,
-    reason,
-    documentEvents: chain.length,
-    redirectResponses: redirectResponses.length,
-    repeatedUrls,
-    repeatedHosts,
-    repeatedPaths,
-    ...(firstUrl ? { firstUrl } : {}),
-    ...(lastUrl ? { lastUrl } : {}),
-    ...(finalAttempted ? { finalAttemptedUrl: finalAttempted } : {}),
-    sanitizedQueryKeys,
-    redactedQueryKeys,
-  }
-
-  return {
-    schema: "wp-codebox/browser-redirect-diagnostics/v1",
-    version: 1,
-    capturedAt: now(),
-    status: "captured",
-    classification,
-    reason,
-    ...(error ? { error: { name: error.name, message: sanitizeBrowserRedirectMessage(error.message) } } : {}),
-    chain,
-    summary,
-  }
-}
-
-function browserRedirectDiagnosticsChainEntry(record: BrowserProbeNetworkRecord): BrowserRedirectDiagnosticsChainEntry {
-  const parsed = parseBrowserRedirectUrl(record.url)
-  return {
-    url: browserRedirectSafeUrl(record.url),
-    method: record.method,
-    ...(typeof record.status === "number" ? { status: record.status } : {}),
-    ...(record.statusText ? { statusText: record.statusText } : {}),
-    timestamp: record.timestamp,
-    ...(parsed ? { host: parsed.host, path: parsed.pathname } : {}),
-    queryKeys: parsed?.queryKeys ?? [],
-    redactedQueryKeys: parsed?.redactedQueryKeys ?? [],
-  }
-}
-
-function browserRedirectSafeUrl(value: string): string {
-  if (/^data:/i.test(value)) {
-    return "data:[redacted]"
-  }
-  const parsed = parseBrowserRedirectUrl(value)
-  if (!parsed) {
-    return value
-  }
-  const search = parsed.queryKeys.length > 0
-    ? `?${parsed.queryKeys.map((key) => `${encodeURIComponent(key)}=[redacted]`).join("&")}`
-    : ""
-  return `${parsed.origin}${parsed.pathname}${search}${parsed.hash ? "#[redacted]" : ""}`
-}
-
-function parseBrowserRedirectUrl(value: string): { origin: string; host: string; pathname: string; hash: string; queryKeys: string[]; redactedQueryKeys: string[] } | undefined {
-  try {
-    const url = new URL(value)
-    const queryKeys = [...new Set([...url.searchParams.keys()])].sort()
-    return {
-      origin: url.origin,
-      host: url.host,
-      pathname: url.pathname || "/",
-      hash: url.hash,
-      queryKeys,
-      redactedQueryKeys: queryKeys.filter(isSensitiveBrowserRedirectQueryKey),
-    }
-  } catch {
-    return undefined
-  }
-}
-
-function isSensitiveBrowserRedirectQueryKey(key: string): boolean {
-  const tokens = key.toLowerCase().split(/[^a-z0-9]+/).filter(Boolean)
-  return tokens.some((token) => ["auth", "bearer", "code", "cookie", "credential", "key", "login", "nonce", "pass", "password", "secret", "session", "state", "token"].includes(token))
-}
-
-function repeatedBrowserRedirectValues<Key extends "url" | "host" | "path">(values: string[], key: Key): Array<Record<Key, string> & { count: number }> {
-  const counts = new Map<string, number>()
-  for (const value of values) {
-    counts.set(value, (counts.get(value) ?? 0) + 1)
-  }
-  return [...counts.entries()]
-    .filter(([, count]) => count > 1)
-    .sort(([leftValue, leftCount], [rightValue, rightCount]) => rightCount - leftCount || leftValue.localeCompare(rightValue))
-    .map(([value, count]) => ({ [key]: value, count }) as Record<Key, string> & { count: number })
-}
-
-function extractBrowserNavigationUrl(message: string): string | undefined {
-  return message.match(/\bat\s+(https?:\/\/\S+)/i)?.[1]?.replace(/[),.]+$/, "")
-}
-
-function sanitizeBrowserRedirectMessage(message: string): string {
-  return message.replace(/https?:\/\/[^\s"')]+/gi, (url) => browserRedirectSafeUrl(url.replace(/[),.]+$/, "")))
-}
-
-interface BrowserWordPressDiagnosticRecord {
-  schema: "wp-codebox/browser-wordpress-diagnostic-record/v1"
-  classification: "php-fatal" | "http-5xx-status" | "http-response-code-5xx"
-  severity: "error"
-  errorType?: number
-  message: string
-  file?: string
-  line?: number
-  status?: number
-  statusHeader?: string
-  requestUri?: string
-  backtrace?: Array<{ file?: string; line?: number; function?: string; class?: string; type?: string }>
-  capturedAt: string
-}
-
-interface BrowserWordPressDiagnosticsArtifact {
-  schema: "wp-codebox/browser-wordpress-diagnostics/v1"
-  version: 1
-  capturedAt: string
-  status: BrowserWordPressDiagnosticsSummary["status"]
-  document5xxResponses: Array<{ url: string; status: number; statusText?: string; responseTextPreview?: string; responseTextSha256?: string; responseTextTruncated?: boolean }>
-  diagnostics: BrowserWordPressDiagnosticRecord[]
-  summary: BrowserWordPressDiagnosticsSummary
-}
-
-const BROWSER_WORDPRESS_DIAGNOSTICS_LOG = "/wordpress/wp-content/wp-codebox-browser-diagnostics.jsonl"
-const BROWSER_WORDPRESS_DIAGNOSTICS_MU_PLUGIN = "/wordpress/wp-content/mu-plugins/000-wp-codebox-browser-diagnostics.php"
-const BROWSER_WORDPRESS_DIAGNOSTICS_PLUGIN = phpBrowserWordPressDiagnosticsPlugin()
-
-async function installBrowserWordPressDiagnostics(
-  runPlaygroundCommand: ((command: string, server: PlaygroundCliServer, options: { code: string } | { scriptPath: string }) => Promise<PlaygroundRunResponse>) | undefined,
-  server: PlaygroundCliServer,
-): Promise<boolean> {
-  if (runPlaygroundCommand) {
-    try {
-      const response = await runPlaygroundCommand("wordpress.browser-diagnostics-setup", server, {
-        code: `<?php
-$directory = '/wordpress/wp-content/mu-plugins';
-if (!is_dir($directory)) {
-    mkdir($directory, 0777, true);
-}
-file_put_contents(${JSON.stringify(BROWSER_WORDPRESS_DIAGNOSTICS_MU_PLUGIN)}, base64_decode(${JSON.stringify(Buffer.from(BROWSER_WORDPRESS_DIAGNOSTICS_PLUGIN, "utf8").toString("base64"))}));
-file_put_contents(${JSON.stringify(BROWSER_WORDPRESS_DIAGNOSTICS_LOG)}, '');
-`,
-      })
-      assertPlaygroundResponseOk("wordpress.browser-diagnostics-setup", response)
-      return true
-    } catch {
-      // Browser diagnostics are best-effort; preserve the browser command outcome.
-    }
-  }
-
-  if (!server.playground.writeFile) {
-    return false
-  }
-
-  try {
-    await server.playground.writeFile(BROWSER_WORDPRESS_DIAGNOSTICS_MU_PLUGIN, BROWSER_WORDPRESS_DIAGNOSTICS_PLUGIN)
-    await server.playground.writeFile(BROWSER_WORDPRESS_DIAGNOSTICS_LOG, "")
-    return true
-  } catch {
-    return false
-  }
-}
-
-async function browserWordPressDiagnosticsArtifact({
-  artifactPath,
-  network,
-  ready,
-  server,
-}: {
-  artifactPath: string
-  network: BrowserProbeNetworkRecord[]
-  ready: boolean
-  server: PlaygroundCliServer
-}): Promise<BrowserWordPressDiagnosticsArtifact | undefined> {
-  const document5xxResponses = network
-    .filter((record) => record.type === "response" && record.resourceType === "document" && typeof record.status === "number" && record.status >= 500 && record.status < 600)
-    .map((record) => ({
-      url: browserRedirectSafeUrl(safeBrowserProbeUrl(record.url) ?? record.url),
-      status: record.status as number,
-      ...(record.statusText ? { statusText: record.statusText } : {}),
-      ...(record.responseTextPreview ? { responseTextPreview: record.responseTextPreview } : {}),
-      ...(record.responseTextSha256 ? { responseTextSha256: record.responseTextSha256 } : {}),
-      ...(typeof record.responseTextTruncated === "boolean" ? { responseTextTruncated: record.responseTextTruncated } : {}),
-    }))
-
-  if (document5xxResponses.length === 0) {
-    return undefined
-  }
-
-  const diagnostics = ready ? await readBrowserWordPressDiagnostics(server) : []
-  const fatalErrors = diagnostics.filter((diagnostic) => diagnostic.classification === "php-fatal").length
-  const classifications = [...new Set(diagnostics.map((diagnostic) => diagnostic.classification))].sort()
-  const status: BrowserWordPressDiagnosticsSummary["status"] = !ready
-    ? "unavailable"
-    : diagnostics.length > 0 ? "captured" : "clean"
-  const summary: BrowserWordPressDiagnosticsSummary = {
-    status,
-    artifact: artifactPath,
-    document5xxResponses: document5xxResponses.length,
-    diagnostics: diagnostics.length,
-    fatalErrors,
-    classifications,
-  }
-
-  return {
-    schema: "wp-codebox/browser-wordpress-diagnostics/v1",
-    version: 1,
-    capturedAt: now(),
-    status,
-    document5xxResponses,
-    diagnostics,
-    summary,
-  }
-}
-
-async function readBrowserWordPressDiagnostics(server: PlaygroundCliServer): Promise<BrowserWordPressDiagnosticRecord[]> {
-  if (!server.playground.readFileAsText) {
-    return []
-  }
-
-  let contents = ""
-  try {
-    contents = await server.playground.readFileAsText(BROWSER_WORDPRESS_DIAGNOSTICS_LOG)
-  } catch {
-    return []
-  }
-
-  return contents
-    .split(/\r?\n/)
-    .map((line) => line.trim())
-    .filter(Boolean)
-    .map(parseBrowserWordPressDiagnosticRecord)
-    .filter((record): record is BrowserWordPressDiagnosticRecord => Boolean(record))
-}
-
-function parseBrowserWordPressDiagnosticRecord(line: string): BrowserWordPressDiagnosticRecord | undefined {
-  let parsed: unknown
-  try {
-    parsed = JSON.parse(line)
-  } catch {
-    return undefined
-  }
-
-  if (!parsed || typeof parsed !== "object" || Array.isArray(parsed)) {
-    return undefined
-  }
-
-  const record = parsed as Record<string, unknown>
-  if (record.schema !== "wp-codebox/browser-wordpress-diagnostic-record/v1" || !isBrowserWordPressDiagnosticClassification(record.classification)) {
-    return undefined
-  }
-
-  const classification = record.classification
-
-  return {
-    schema: "wp-codebox/browser-wordpress-diagnostic-record/v1",
-    classification,
-    severity: "error",
-    ...(typeof record.errorType === "number" && Number.isFinite(record.errorType) ? { errorType: record.errorType } : {}),
-    message: sanitizeBrowserWordPressDiagnosticString(typeof record.message === "string" ? record.message : ""),
-    ...(typeof record.file === "string" && record.file.length > 0 ? { file: sanitizeBrowserWordPressDiagnosticString(record.file) } : {}),
-    ...(typeof record.line === "number" && Number.isFinite(record.line) ? { line: record.line } : {}),
-    ...(typeof record.status === "number" && Number.isFinite(record.status) ? { status: record.status } : {}),
-    ...(typeof record.statusHeader === "string" && record.statusHeader.length > 0 ? { statusHeader: sanitizeBrowserWordPressDiagnosticString(record.statusHeader) } : {}),
-    ...(typeof record.requestUri === "string" && record.requestUri.length > 0 ? { requestUri: sanitizeBrowserWordPressDiagnosticRequestUri(record.requestUri) } : {}),
-    ...(Array.isArray(record.backtrace) ? { backtrace: sanitizeBrowserWordPressDiagnosticBacktrace(record.backtrace) } : {}),
-    capturedAt: typeof record.capturedAt === "string" ? record.capturedAt : now(),
-  }
-}
-
-function isBrowserWordPressDiagnosticClassification(value: unknown): value is BrowserWordPressDiagnosticRecord["classification"] {
-  return value === "php-fatal" || value === "http-5xx-status" || value === "http-response-code-5xx"
-}
-
-function sanitizeBrowserWordPressDiagnosticString(value: string): string {
-  return value
-    .replace(/https?:\/\/[^\s"'<>]+/gi, (url) => browserRedirectSafeUrl(url))
-    .replace(/([?&][^=&#\s"'<>]+)=([^&#\s"'<>]+)/g, "$1=[redacted]")
-    .replace(/((?:access[_-]?token|auth|bearer|code|cookie|credential|key|login|nonce|pass|password|secret|session|state|token)["'\s:=]+)[^\s"'<>]+/gi, "$1[redacted]")
-}
-
-function sanitizeBrowserWordPressDiagnosticRequestUri(value: string): string {
-  try {
-    const parsed = new URL(value, "http://wp-codebox.local")
-    const queryKeys = [...new Set([...parsed.searchParams.keys()])].sort()
-    const query = queryKeys.length > 0 ? `?${queryKeys.map((key) => `${encodeURIComponent(key)}=[redacted]`).join("&")}` : ""
-    return `${parsed.pathname || "/"}${query}${parsed.hash ? "#[redacted]" : ""}`
-  } catch {
-    return sanitizeBrowserWordPressDiagnosticString(value)
-  }
-}
-
-function sanitizeBrowserWordPressDiagnosticBacktrace(value: unknown[]): BrowserWordPressDiagnosticRecord["backtrace"] {
-  return value.flatMap((entry) => {
-    if (!entry || typeof entry !== "object" || Array.isArray(entry)) {
-      return []
-    }
-    const frame = entry as Record<string, unknown>
-    return [{
-      ...(typeof frame.file === "string" && frame.file.length > 0 ? { file: sanitizeBrowserWordPressDiagnosticString(frame.file) } : {}),
-      ...(typeof frame.line === "number" && Number.isFinite(frame.line) ? { line: frame.line } : {}),
-      ...(typeof frame.function === "string" && frame.function.length > 0 ? { function: sanitizeBrowserWordPressDiagnosticString(frame.function) } : {}),
-      ...(typeof frame.class === "string" && frame.class.length > 0 ? { class: sanitizeBrowserWordPressDiagnosticString(frame.class) } : {}),
-      ...(typeof frame.type === "string" && frame.type.length > 0 ? { type: sanitizeBrowserWordPressDiagnosticString(frame.type) } : {}),
-    }]
-  }).slice(0, 12)
-}
 
 export async function runHtmlCaptureCommand(input: {
   artifactRoot: string
@@ -3548,389 +2015,6 @@ function summarizeEditorState(target: ReturnType<typeof editorOpenTargetFromArgs
   }
 }
 
-async function installWordPressAdminAuthCookies({
-  cookieUrls,
-  command,
-  page,
-  runPlaygroundCommand,
-  runtimeSpec,
-  server,
-  userId,
-}: {
-  cookieUrls?: string[]
-  command: string
-  page: import("playwright").Page
-  runPlaygroundCommand?: (command: string, server: PlaygroundCliServer, options: { code: string } | { scriptPath: string }) => Promise<PlaygroundRunResponse>
-  runtimeSpec?: RuntimeCreateSpec
-  server: PlaygroundCliServer
-  userId: number
-}): Promise<BrowserProbeAuthSummary> {
-  if (!runPlaygroundCommand) {
-    throw new Error(`${command} auth=wordpress-admin requires Playground PHP command support`)
-  }
-  if (!runtimeSpec) {
-    throw new Error(`${command} auth=wordpress-admin requires a runtime spec`)
-  }
-
-  const authCommand = `${command}.auth`
-  const urls = uniqueBrowserAuthCookieUrls(cookieUrls ?? [server.serverUrl])
-  const response = await runPlaygroundCommand(authCommand, server, { code: bootstrapPhpCode(runtimeSpec, wordpressAdminAuthCookiePhpCode(urls, userId), []) })
-  assertPlaygroundResponseOk(authCommand, response)
-  const cookies = JSON.parse(cleanWpCliOutput(response.text)) as Array<{ name?: string; value?: string; domain?: string; path?: string; expires?: number; httpOnly?: boolean; secure?: boolean; sameSite?: "Lax" }>
-  await page.context().addCookies(cookies.map((cookie) => ({
-    name: String(cookie.name ?? ""),
-    value: String(cookie.value ?? ""),
-    domain: String(cookie.domain ?? new URL(server.serverUrl).hostname),
-    path: typeof cookie.path === "string" && cookie.path.length > 0 ? cookie.path : "/",
-    expires: typeof cookie.expires === "number" ? cookie.expires : Math.floor(Date.now() / 1000) + 3600,
-    httpOnly: cookie.httpOnly !== false,
-    secure: cookie.secure === true,
-    sameSite: cookie.sameSite ?? "Lax",
-  })))
-
-  return { mode: "wordpress-admin", userId, cookieCount: cookies.length, cookieHosts: browserAuthCookieHostSummary(cookies) }
-}
-
-export function wordpressAdminAuthCookiePhpCode(browserUrls: string[], userId: number): string {
-  return `
-$user_id = ${JSON.stringify(userId)};
-$user = get_user_by( 'id', $user_id );
-if ( ! $user ) {
-    throw new RuntimeException( 'Browser auth requires the requested WordPress user to exist.' );
-}
-wp_set_current_user( $user_id );
-$expiration = time() + HOUR_IN_SECONDS;
-$token = '';
-if ( class_exists( 'WP_Session_Tokens' ) ) {
-    $token = WP_Session_Tokens::get_instance( $user_id )->create( $expiration );
-}
-$browser_urls = ${JSON.stringify(browserUrls)};
-$cookies = array();
-foreach ( $browser_urls as $browser_url ) {
-    $browser_host = wp_parse_url( $browser_url, PHP_URL_HOST );
-    if ( ! $browser_host ) {
-        continue;
-    }
-    $secure = 'https' === wp_parse_url( $browser_url, PHP_URL_SCHEME );
-    foreach ( array( array( AUTH_COOKIE, 'auth', false ), array( SECURE_AUTH_COOKIE, 'secure_auth', true ) ) as $admin_cookie ) {
-        $cookies[] = array(
-            'name'     => $admin_cookie[0],
-            'value'    => wp_generate_auth_cookie( $user_id, $expiration, $admin_cookie[1], $token ),
-            'domain'   => $browser_host,
-            'path'     => defined( 'ADMIN_COOKIE_PATH' ) && ADMIN_COOKIE_PATH ? ADMIN_COOKIE_PATH : '/wp-admin',
-            'expires'  => $expiration,
-            'httpOnly' => true,
-            'secure'   => $admin_cookie[2],
-            'sameSite' => 'Lax',
-        );
-    }
-    $logged_in_cookie = array(
-        'name'     => LOGGED_IN_COOKIE,
-        'value'    => wp_generate_auth_cookie( $user_id, $expiration, 'logged_in', $token ),
-        'domain'   => $browser_host,
-        'path'     => defined( 'COOKIEPATH' ) && COOKIEPATH ? COOKIEPATH : '/',
-        'expires'  => $expiration,
-        'httpOnly' => true,
-        'secure'   => $secure,
-        'sameSite' => 'Lax',
-    );
-    $cookies[] = $logged_in_cookie;
-    if ( defined( 'SITECOOKIEPATH' ) && SITECOOKIEPATH && SITECOOKIEPATH !== COOKIEPATH ) {
-        $logged_in_cookie['path'] = SITECOOKIEPATH;
-        $cookies[] = $logged_in_cookie;
-    }
-}
-echo wp_json_encode( $cookies );
-`
-}
-
-function browserActionTargetUrls(steps: BrowserInteractionStep[], effectiveOrigin: string, fallbackUrl: string): string[] {
-  const urls = steps
-    .filter((step) => step.kind === "navigate" && typeof step.url === "string" && step.url.trim().length > 0)
-    .map((step) => resolveBrowserPreviewUrl(String(step.url), effectiveOrigin))
-  return urls.length > 0 ? urls : [fallbackUrl]
-}
-
-function uniqueBrowserAuthCookieUrls(urls: string[]): string[] {
-  const unique = new Map<string, string>()
-  for (const url of urls) {
-    try {
-      const parsed = new URL(url)
-      unique.set(`${parsed.protocol}//${normalizeBrowserCookieHost(parsed.hostname)}`, `${parsed.protocol}//${parsed.hostname}/`)
-    } catch {
-      // Ignore invalid cookie URL inputs; callers still include the local server URL.
-    }
-  }
-  return [...unique.values()]
-}
-
-function normalizeBrowserCookieHost(host: string): string {
-  return host.trim().toLowerCase().replace(/:\d+$/, "")
-}
-
-function browserAuthCookieHostSummary(cookies: Array<{ domain?: string }>): Array<{ host: string; cookieCount: number }> {
-  const counts = new Map<string, number>()
-  for (const cookie of cookies) {
-    const host = normalizeBrowserCookieHost(String(cookie.domain ?? ""))
-    if (!host) continue
-    counts.set(host, (counts.get(host) ?? 0) + 1)
-  }
-  return [...counts.entries()].sort(([left], [right]) => left.localeCompare(right)).map(([host, cookieCount]) => ({ host, cookieCount }))
-}
-
-function browserAuthRequest(args: string[]): { userId: number } | undefined {
-  const auth = argValue(args, "auth")?.trim()
-  if (!auth) {
-    return undefined
-  }
-  if (auth !== "wordpress-admin") {
-    throw new Error(`Browser auth supports wordpress-admin: ${auth}`)
-  }
-  return { userId: positiveIntegerArg(args, "auth-user-id", 1) }
-}
-
-async function browserStorageStateImportFromArgs(args: string[], command: string): Promise<BrowserStorageStateImport | undefined> {
-  const raw = argValue(args, "storage-state")?.trim()
-  if (!raw) {
-    return undefined
-  }
-
-  const source = raw.startsWith("@") ? "file" : "inline"
-  const text = source === "file" ? await readFile(resolveCommandPath(raw.slice(1)), "utf8") : raw
-  let payload: unknown
-  try {
-    payload = JSON.parse(text)
-  } catch (error) {
-    throw new BrowserStorageStateImportError(`${command} storage-state must be valid JSON`, {
-      status: "error",
-      source,
-      cookieCount: 0,
-      cookieHosts: [],
-      originCount: 0,
-      diagnostics: [{ code: "storage-state-json-invalid", severity: "error", message: error instanceof Error ? error.message : String(error) }],
-    })
-  }
-
-  const normalized = normalizeBrowserStorageStatePayload(payload, source)
-  if (normalized.summary.status !== "ready") {
-    throw new BrowserStorageStateImportError(`${command} storage-state is unsupported`, normalized.summary)
-  }
-  return normalized
-}
-
-function browserStorageStateAuthSummary(summary: BrowserStorageStateImportSummary): BrowserProbeAuthSummary {
-  return {
-    mode: "storage-state",
-    storageState: summary,
-    cookieCount: summary.cookieCount,
-    cookieHosts: summary.cookieHosts,
-  }
-}
-
-class BrowserStorageStateImportError extends Error {
-  constructor(message: string, readonly storageState: BrowserStorageStateImportSummary) {
-    super(message)
-    this.name = "BrowserStorageStateImportError"
-  }
-
-  toJSON(): { name: string; message: string; storageState: BrowserStorageStateImportSummary } {
-    return { name: this.name, message: this.message, storageState: this.storageState }
-  }
-}
-
-async function fileSha256(path: string): Promise<string> {
-  return sha256(await readFile(path))
-}
-
-type BrowserProbeProgressSource = "navigation" | "network" | "console" | "pageerror" | "checkpoint" | "script" | "duration" | "probe-error"
-
-class BrowserProbeTerminalFailureError extends Error {
-  readonly code = "browser-probe-terminal-failure"
-
-  constructor(readonly failure: { message: string; reason?: string; details?: unknown; timestamp: string }) {
-    super(`Browser probe reported a terminal failure: ${failure.message}`)
-    this.name = "BrowserProbeTerminalFailureError"
-  }
-}
-
-class BrowserProbeStallError extends Error {
-  readonly code = "browser-probe-stalled"
-
-  constructor(readonly idleMs: number, readonly stallTimeoutMs: number, readonly lastProgressSource: BrowserProbeProgressSource, readonly lastCheckpoint?: BrowserProbeScriptCheckpoint) {
-    super(`Browser probe stalled after ${idleMs}ms without progress; last progress source was ${lastProgressSource}${lastCheckpoint ? ` (${lastCheckpoint.name})` : ""}`)
-    this.name = "BrowserProbeStallError"
-  }
-}
-
-function createBrowserProbeProgressTracker(startedAt: string, stallTimeoutMs: number): {
-  mark(source: BrowserProbeProgressSource, timestamp?: string, checkpoint?: BrowserProbeScriptCheckpoint): void
-  fail(source: BrowserProbeProgressSource, error: Error): void
-  terminalFailure(failure: { message: string; reason?: string; details?: unknown; timestamp: string }): void
-  lastProgressElapsedMs(): number
-  summary(): {
-    status: "active" | "failed" | "stalled"
-    startedAt: string
-    lastProgressAt: string
-    lastProgressSource: BrowserProbeProgressSource
-    idleMs: number
-    stallTimeoutMs?: number
-    lastCheckpoint?: BrowserProbeScriptCheckpoint
-    terminalFailure?: { message: string; reason?: string; details?: unknown; timestamp: string }
-  }
-} {
-  let status: "active" | "failed" | "stalled" = "active"
-  let lastProgressAt = startedAt
-  let lastProgressSource: BrowserProbeProgressSource = "navigation"
-  let lastCheckpoint: BrowserProbeScriptCheckpoint | undefined
-  let terminalFailure: { message: string; reason?: string; details?: unknown; timestamp: string } | undefined
-
-  return {
-    mark(source, timestamp = now(), checkpoint) {
-      lastProgressAt = timestamp
-      lastProgressSource = source
-      if (checkpoint) {
-        lastCheckpoint = checkpoint
-      }
-    },
-    fail(source, error) {
-      lastProgressAt = now()
-      lastProgressSource = source
-      status = error instanceof BrowserProbeStallError ? "stalled" : "failed"
-      if (error instanceof BrowserProbeTerminalFailureError) {
-        terminalFailure = error.failure
-      }
-    },
-    terminalFailure(failure) {
-      status = "failed"
-      terminalFailure = failure
-      lastProgressAt = failure.timestamp
-      lastProgressSource = "probe-error"
-    },
-    lastProgressElapsedMs() {
-      return Math.max(0, Date.now() - Date.parse(lastProgressAt))
-    },
-    summary() {
-      return {
-        status,
-        startedAt,
-        lastProgressAt,
-        lastProgressSource,
-        idleMs: this.lastProgressElapsedMs(),
-        ...(stallTimeoutMs > 0 ? { stallTimeoutMs } : {}),
-        ...(lastCheckpoint ? { lastCheckpoint } : {}),
-        ...(terminalFailure ? { terminalFailure } : {}),
-      }
-    },
-  }
-}
-
-async function withBrowserProbeLiveness<T>(page: import("playwright").Page, progress: ReturnType<typeof createBrowserProbeProgressTracker>, failFast: boolean, operation: Promise<T>, policy: Required<BrowserCommandLivenessPolicy>, phase: string): Promise<T> {
-  const result = await withBrowserCommandLiveness({
-    command: "wordpress.browser-probe",
-    phase,
-    operation,
-    policy,
-    idle: () => {
-      const summary = progress.summary()
-      return { idleMs: summary.idleMs, lastProgressSource: summary.lastProgressSource }
-    },
-    poll: async () => {
-      try {
-        const state = await page.evaluate(() => {
-          const probe = (globalThis as typeof globalThis & {
-            __wpCodeboxBrowserProbe?: {
-              checkpoints?: Array<{ name?: unknown; metadata?: unknown; timestamp?: unknown }>
-              terminalFailure?: { message?: unknown; reason?: unknown; details?: unknown; timestamp?: unknown }
-            }
-          }).__wpCodeboxBrowserProbe
-          const checkpoints = Array.isArray(probe?.checkpoints) ? probe.checkpoints : []
-          const latestCheckpoint = [...checkpoints].reverse().find((checkpoint) => typeof checkpoint.timestamp === "string")
-          const latestCheckpointTimestamp = typeof latestCheckpoint?.timestamp === "string" ? latestCheckpoint.timestamp : undefined
-          const checkpoint = latestCheckpoint && latestCheckpointTimestamp ? {
-            name: typeof latestCheckpoint.name === "string" ? latestCheckpoint.name : "checkpoint",
-            metadata: latestCheckpoint.metadata,
-            timestamp: latestCheckpointTimestamp,
-          } : undefined
-          const failure = probe?.terminalFailure
-          return {
-            checkpoint,
-            terminalFailure: failure && typeof failure.message === "string" ? {
-              message: failure.message,
-              reason: typeof failure.reason === "string" ? failure.reason : undefined,
-              details: failure.details,
-              timestamp: typeof failure.timestamp === "string" ? failure.timestamp : new Date().toISOString(),
-            } : undefined,
-          }
-        })
-        if (state.checkpoint) {
-          progress.mark("checkpoint", state.checkpoint.timestamp, state.checkpoint)
-        }
-        if (state.terminalFailure) {
-          progress.terminalFailure(state.terminalFailure)
-          throw new BrowserProbeTerminalFailureError(state.terminalFailure)
-        }
-      } catch (error) {
-        if (error instanceof BrowserProbeTerminalFailureError) {
-          throw error
-        }
-        // The page may be navigating or already closed; the outer operation remains authoritative.
-      }
-    },
-    onTimeout: async () => {
-      await page.close().catch(() => undefined)
-    },
-  }).catch((error) => {
-    if (isBrowserCommandLivenessError(error) && error.code === "browser-command-idle-timeout") {
-      const summary = progress.summary()
-      throw new BrowserProbeStallError(summary.idleMs, policy.idleTimeoutMs, summary.lastProgressSource, summary.lastCheckpoint)
-    }
-    throw error
-  })
-  const terminalFailure = failFast ? await browserProbeTerminalFailure(page) : undefined
-  if (terminalFailure) {
-    progress.terminalFailure(terminalFailure)
-    throw new BrowserProbeTerminalFailureError(terminalFailure)
-  }
-  return result
-}
-
-function livenessRemainingWallTimeMs(startedAtMs: number, totalTimeoutMs: number): number {
-  if (totalTimeoutMs <= 0) {
-    return browserCommandLivenessPolicy().wallTimeoutMs
-  }
-  return Math.max(1, totalTimeoutMs - (Date.now() - startedAtMs))
-}
-
-function normalizeBrowserProbeScriptCheckpoint(checkpoint: unknown): BrowserProbeScriptCheckpoint | undefined {
-  if (!checkpoint || typeof checkpoint !== "object") {
-    return undefined
-  }
-  const record = checkpoint as { name?: unknown; metadata?: unknown; timestamp?: unknown }
-  return {
-    name: typeof record.name === "string" && record.name.length > 0 ? record.name : "checkpoint",
-    ...(typeof record.metadata !== "undefined" ? { metadata: record.metadata } : {}),
-    timestamp: typeof record.timestamp === "string" ? record.timestamp : now(),
-  }
-}
-
-async function browserProbeTerminalFailure(page: import("playwright").Page): Promise<{ message: string; reason?: string; details?: unknown; timestamp: string } | undefined> {
-  return page.evaluate(() => {
-    const failure = (globalThis as typeof globalThis & {
-      __wpCodeboxBrowserProbe?: {
-        terminalFailure?: { message?: unknown; reason?: unknown; details?: unknown; timestamp?: unknown }
-      }
-    }).__wpCodeboxBrowserProbe?.terminalFailure
-    if (!failure || typeof failure.message !== "string") {
-      return undefined
-    }
-    return {
-      message: failure.message,
-      reason: typeof failure.reason === "string" ? failure.reason : undefined,
-      details: failure.details,
-      timestamp: typeof failure.timestamp === "string" ? failure.timestamp : new Date().toISOString(),
-    }
-  }).catch(() => undefined)
-}
 
 function numberArg(args: string[], name: string, fallback: number): number {
   const raw = argValue(args, name)?.trim()
